+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/model/KtorfitApplicationShort.kt
@@ -22,7 +22,6 @@ data class KtorfitApplicationShort(
     val updatedAt: LocalDateTime,
     @SerialName("category_id") val categoryId: String,
     @SerialName("alias") val alias: String,
-    @SerialName("name") val name: String,
     @SerialName("current_version") val currentVersion: KtorfitCurrentVersionShort,
 ) {
     fun toFapItemShort(category: FapCategory?, target: FlipperTarget): FapItemShort? {
@@ -39,8 +38,8 @@ data class KtorfitApplicationShort(
         return FapItemShort(
             id = id,
             picUrl = currentVersion.iconUrl,
-            description = currentVersion.shortDescription,
-            name = name,
+            shortDescription = currentVersion.shortDescription,
+            name = currentVersion.name,
             category = category,
             screenshots = currentVersion.screenshots.toImmutableList(),
             applicationAlias = alias,